requirements.txt
torch
torchvision
numpy
pandas
matplotlib
scikit-learn
opencv-python
tqdm
pyyaml
albumentations
kagglehub
tensorboard          # Added for tensorboard logging
tensorboardX         # Optional, for advanced logging (if needed)
Pillow               # For image handling (PIL images)